+++ src/components/Sushi.js
@@ -19,7 +19,6 @@ function Sushi({
   return (
     <div className="sushi">
       <div className="plate" id={id} onClick={onPlateClick}>
-        {/* Tell me if this sushi has been eaten! */}
         {isEaten ? null : <img id={id} src={img_url} alt={name} width="100%" />}
       </div>
       <h4 className="sushi-details">
